Reorganize the code for better maintenance
