+++ src/types.ts
@@ -1,0 +1,9 @@
+/**
+ * Error type for OpenAI API errors
+ */
+export interface APIError extends Error {
+	response?: {
+		status: number;
+		data: any;
+	};
+}
